Rename chronostratigraphy folder to stratigraphy
