+++ src/client/src/commons/form/borehole/stratigraphy/chronostratigraphyEditProfile.js
