+++ src/client/src/commons/form/borehole/stratigraphy/chronostratigraphyPanel.js
