+++ src/client/src/commons/form/borehole/stratigraphy/hierarchicalDataEditProfile.js
