+++ src/client/src/commons/form/borehole/stratigraphy/layerCard.js
