+++ src/client/src/commons/form/borehole/stratigraphy/lithologyViewProfile.js
