+++ src/client/src/commons/form/borehole/stratigraphy/lithostratigraphyEditProfile.js
