+++ src/client/src/commons/form/borehole/stratigraphy/lithostratigraphyPanel.js
