+++ src/client/src/commons/form/borehole/stratigraphy/lithostratigraphyViewProfile.js
