+++ src/client/src/commons/form/borehole/stratigraphy/navigationChild.js
