+++ src/client/src/commons/form/borehole/stratigraphy/navigationContainer.js
