+++ src/client/src/commons/form/borehole/stratigraphy/navigationLens.js
