+++ src/client/src/commons/form/borehole/stratigraphy/scale.js
