+++ src/client/src/commons/form/borehole/stratigraphy/stratigraphySelection.js
